accounts/utils.py
import jwt
from django.http    import JsonResponse

from accounts.models   import User
from django.conf import settings


#로그인한 유저인지 확인
def login_decorator(func):
    def wrapper(self, request, *args, **kwargs):
        try:
            access_token = request.headers.get('Authorization', None)
            payload      = jwt.decode(access_token, settings.SECRET_KEY, algorithms='HS256')
            user         = User.objects.get(id=payload['id'])
            request.user = user
            
        except jwt.DecodeError:
            return JsonResponse({'MESSAGE' : 'INVALID_TOKEN'}, status=400)
            
        except User.DoesNotExist:
            return JsonResponse({'MESSAGE' : 'INVALID_USER'}, status=400)
        
        return func(self, request, *args, **kwargs)
    
    return wrapper 